+++ common/src/main/java/com/evandev/music_tweaks/mixin/SoundManagerAccessor.java
@@ -1,0 +1,15 @@
+package com.evandev.music_tweaks.mixin;
+
+import net.minecraft.client.sounds.SoundManager;
+import net.minecraft.client.sounds.WeighedSoundEvents;
+import net.minecraft.resources.ResourceLocation;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+import java.util.Map;
+
+@Mixin(SoundManager.class)
+public interface SoundManagerAccessor {
+    @Accessor("registry")
+    Map<ResourceLocation, WeighedSoundEvents> getRegistry();
+}